+++ JejuPlanner/src/main/java/com/gteam/planner/controller/MemberController.java
@@ -9,6 +9,7 @@ import javax.servlet.http.HttpSession;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -28,12 +29,10 @@ import com.gteam.planner.service.MemberService;
 public class MemberController {
 	
 	private static final Logger log = LoggerFactory.getLogger(MemberController.class);
-	private final MemberService service;
 	
+	@Autowired
+	private MemberService service;
 	
-	public MemberController(MemberService service) {
-		this.service = service;
-	}
 	
 	// localhost:9090/ GET 요청 시 login.jsp로 던져줌
 	@RequestMapping(value = "/", method = RequestMethod.GET)
@@ -93,7 +92,7 @@ public class MemberController {
 		}else {
 			//DB에서 회원정보 존재 시 view단에 회원정보(아이디, 패스워드 반환)
 			session.setAttribute("member", login);
-			return "redirect:/plan/schedulePlan";
+			return "redirect:/plan/write";
 		}
 		
 		
@@ -107,27 +106,27 @@ public class MemberController {
 		
 		return "redirect:/";
 	}
-	//로그인 후 일정 만들기 화면으로 이동
-	@RequestMapping(value="/plan/schedulePlan", method = RequestMethod.GET)
-	public void schedulePlanning(Model model) throws Exception{
-		if (!model.containsAttribute("cart")) {
-			model.addAttribute("cart", new ArrayList<PlanVO>());
-		}
-		
-	}
-	
-	@RequestMapping(value="/plan/schedulePlan/add", method=RequestMethod.POST)
-	public String addSchedule(
-			//입력 파라미터인 name, price, cnt가 goods에 일괄 바인딩 됨
-            //goods라는 이름의 Model로 저장(Request Scope)
-            @ModelAttribute PlanVO vo, 
-            //세션에서 cart라는 이름의 Model을 가져와 cart에 할당
-            @ModelAttribute("cart") List<PlanVO> cart) {
-      // cart라는 Model(ArrayList)에 사용자가 입력한 항목들을 가진 goods 객체를 Add
-		log.info("Modal Plan Add");
-      cart.add(vo);
-
-      //localhost:port/cart로 요청을 보냄
-      return "redirect:/plan/schedulePlan";
-	}
+//	//로그인 후 일정 만들기 화면으로 이동
+//	@RequestMapping(value="/plan/schedulePlan", method = RequestMethod.GET)
+//	public void schedulePlanning(Model model) throws Exception{
+//		if (!model.containsAttribute("cart")) {
+//			model.addAttribute("cart", new ArrayList<PlanVO>());
+//		}
+//		
+//	}
+//	
+//	@RequestMapping(value="/plan/schedulePlan/add", method=RequestMethod.POST)
+//	public String addSchedule(
+//			//입력 파라미터인 name, price, cnt가 goods에 일괄 바인딩 됨
+//            //goods라는 이름의 Model로 저장(Request Scope)
+//            @ModelAttribute PlanVO vo, 
+//            //세션에서 cart라는 이름의 Model을 가져와 cart에 할당
+//            @ModelAttribute("cart") List<PlanVO> cart) {
+//      // cart라는 Model(ArrayList)에 사용자가 입력한 항목들을 가진 goods 객체를 Add
+//		log.info("Modal Plan Add");
+//      cart.add(vo);
+//
+//      //localhost:port/cart로 요청을 보냄
+//      return "redirect:/plan/schedulePlan";
+//	}
 }
